LangChain-Chatbot/backend/requirements.txt
fastapi
uvicorn
python-dotenv
pypdf
sentence-transformers
faiss-cpu
numpy
ollama
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
python-multipart
email-validator
